requirements.txt
absl-py==2.0.0
appdirs==1.4.4
av==10.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
cmake==3.26.4
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.44.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.32.0
importlib-resources==6.1.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
onedrivedownloader==1.1.3
packaging==23.2
pathtools==0.1.2
Pillow==10.0.0
pip==23.1.2
protobuf==4.23.3
psutil==5.9.5
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0
quadprog==0.1.11
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
sentry-sdk==1.26.0
setproctitle==1.3.2
setuptools==67.8.0
six==1.16.0
smmap==5.0.0
sympy==1.12
tifffile==2023.9.26
torch==2.0.0+cu118
torchvision==0.15.0+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.3
wandb==0.15.4
wheel==0.38.4
zipp==3.17.0
